requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
spacy>=3.6.0
faiss-gpu>=1.7.4
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
moviepy>=1.0.3
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl 